Refactor makefile tools
